Issue/13 refactor: restructure directory layout (#13)
