+++ src/main/java/decideProgram/Decide.java
@@ -1,4 +1,4 @@
-package assignment1;
+package decideProgram;
 
 public class Decide {
