+++ src/main/java/decideProgram/LIC.java
@@ -1,4 +1,4 @@
-package assignment1;
+package decideProgram;
 
 public class LIC {
